# docs/swagger.yaml
basePath: /api
definitions:
...
  models.Account:
    properties:
      balance:
        description: Поточний кешований баланс
        type: integer
      bank_name:
        type: string
      card_number:
        description: Банківські деталі
        type: string
      card_type:
        description: Credit/Debit
        type: string
      color:
        type: string
      created_at:
        type: integer
      currency:
        type: string
      deleted_at:
        description: Індекс для soft delete
        type: integer
      family_id:
        type: string
      goal:
        $ref: '#/definitions/models.Goal'
      goal_id:
        description: Зв'язок з ціллю
        type: string
      id:
        type: string
      initial_balance:
        description: Баланси
        type: integer
      is_archived:
        type: boolean
      is_group:
        type: boolean
      is_synced:
        type: boolean
      name:
        type: string
      parent_id:
        description: Для вкладеності
        type: string
      payment_system:
        description: Visa/Mastercard
        type: string
      storage_type:
        $ref: '#/definitions/models.StorageType'
      storage_type_id:
        description: Для скарбничок (Piggy Bank)
        type: string
      type:
        description: 'Типи: ''card'', ''cash'', ''crypto'', ''piggy_bank'''
        type: string
      updated_at:
        type: integer
      user_id:
        type: string
    type: object
  models.Asset:
    properties:
      address:
        description: Нерухомість
        type: string
      area:
        type: number
      cadastral_num:
        type: string
      created_at:
        type: integer
      currency:
        type: string
      current_price:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      depreciation_type:
        description: Амортизація
        type: string
      documents:
        description: "\U0001F525 ДОДАНО: Документи (PDF та інші)"
        items:
          $ref: '#/definitions/models.AssetDocument'
        type: array
      estimated_life:
        type: integer
      family_id:
        type: string
      id:
        type: string
      initial_mileage:
        type: integer
      initial_value:
        type: integer
      insurance_expiry:
        type: integer
      is_sold:
        description: Продаж
        type: boolean
      is_synced:
        type: boolean
      last_service_date:
        type: integer
      mileage:
        type: integer
      name:
        type: string
      note:
        type: string
      photo:
        description: Медіа
        type: string
      photos:
        items:
          $ref: '#/definitions/models.AssetPhoto'
        type: array
      price:
        description: Фінанси
        type: integer
      purchase_date:
        type: integer
      serial_number:
        description: Деталі
        type: string
      sold_date:
        type: integer
      sold_price:
        type: integer
      type:
        description: '''car'', ''real_estate'', ''appliance'', ''furniture'''
        type: string
      updated_at:
        type: integer
      user_id:
        type: string
      vin_code:
        description: Автоспецифіка
        type: string
      warranty_end:
        type: integer
    type: object
  models.AssetDocument:
    properties:
      asset_id:
        type: string
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      file_type:
        description: Розширення (напр. "pdf")
        type: string
      id:
        type: string
      is_synced:
        type: boolean
      name:
        description: Оригінальна назва файлу (напр. "чек.pdf")
        type: string
      path:
        description: Шлях на сервері
        type: string
      size:
        description: Розмір в байтах (опціонально, але корисно)
        type: integer
      updated_at:
        type: integer
    type: object
  models.AssetPhoto:
    properties:
      asset_id:
        type: string
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      id:
        type: string
      is_synced:
        type: boolean
      path:
        type: string
      updated_at:
        type: integer
    type: object
...
  models.Category:
    properties:
      color:
        type: string
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      family_id:
        type: string
      icon:
        type: string
      id:
        type: string
      is_synced:
        type: boolean
      name:
        type: string
      parent_id:
        type: string
      type:
        description: '''expense'', ''income'''
        type: string
      updated_at:
        type: integer
    type: object
  models.Counterparty:
    properties:
      balances:
        description: Баланси (борги)
        items:
          $ref: '#/definitions/models.CounterpartyBalance'
        type: array
      category:
        $ref: '#/definitions/models.CounterpartyCategory'
      category_id:
        type: string
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      family_id:
        type: string
      icon:
        type: string
      id:
        type: string
      is_group:
        type: boolean
      is_synced:
        type: boolean
      logo:
        type: string
      name:
        type: string
      parent_id:
        type: string
      type:
        description: '''shop'', ''person'', ''other'''
        type: string
      updated_at:
        type: integer
    type: object
  models.CounterpartyBalance:
    properties:
      balance:
        description: + нам винні, - ми винні
        type: integer
      counterparty_id:
        type: string
      currency:
        type: string
    type: object
  models.CounterpartyCategory:
    properties:
      color:
        type: string
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      family_id:
        type: string
      icon:
        type: string
      id:
        type: string
      is_synced:
        type: boolean
      name:
        type: string
      type:
        type: string
      updated_at:
        type: integer
    type: object
...
  models.Family:
    properties:
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      id:
        type: string
      is_synced:
        type: boolean
      name:
        type: string
      updated_at:
        type: integer
    type: object
  models.Goal:
    properties:
      accounts:
        description: Relations
        items:
          $ref: '#/definitions/models.Account'
        type: array
      color:
        type: string
      created_at:
        type: integer
      currency:
        type: string
      current_amount:
        description: Вираховується динамічно
        type: integer
      date_deadline:
        type: integer
      date_start:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      description:
        type: string
      external_link:
        type: string
      family_id:
        type: string
      hidden_from:
        description: ID юзера, від якого приховано
        type: string
      icon:
        type: string
      id:
        type: string
      is_synced:
        type: boolean
      name:
        type: string
      photo_url:
        type: string
      remove_photo:
        type: boolean
      status:
        description: active, reached, failed
        type: string
      target_amount:
        type: integer
      updated_at:
        type: integer
      user_id:
        description: '<-- ДОДАНО: Власник цілі'
        type: string
      visibility:
        description: '<-- ДОДАНО: Налаштування приватності'
        type: string
    type: object
  models.Role:
    properties:
      can_edit_schema:
        type: boolean
      can_manage_users:
        type: boolean
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      description:
        type: string
      id:
        type: string
      is_synced:
        type: boolean
      name:
        type: string
      updated_at:
        type: integer
    type: object
...
  models.StorageType:
    properties:
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      family_id:
        description: Pointer для NULL (системні типи)
        type: string
      icon:
        type: string
      id:
        type: string
      is_synced:
        type: boolean
      is_system:
        type: boolean
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: integer
    type: object
  models.Tag:
    properties:
      color:
        type: string
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      family_id:
        type: string
      id:
        type: string
      is_synced:
        type: boolean
      name:
        type: string
      updated_at:
        type: integer
    type: object
  models.Transaction:
    properties:
      account:
        allOf:
        - $ref: '#/definitions/models.Account'
        description: Relations
      account_id:
        type: string
      amount:
        type: integer
      asset:
        $ref: '#/definitions/models.Asset'
      asset_id:
        description: Зв'язки з модулями
        type: string
      category:
        $ref: '#/definitions/models.Category'
      category_id:
        type: string
      counterparty:
        $ref: '#/definitions/models.Counterparty'
      counterparty_id:
        type: string
      created_at:
        type: integer
      currency:
        type: string
      date:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      external_id:
        description: Зв'язок із зовнішніми інтеграціями
        type: string
      family_id:
        type: string
      id:
        type: string
      is_forgiveness:
        description: Прощення боргу
        type: boolean
      is_synced:
        type: boolean
      items:
        items:
          $ref: '#/definitions/models.TransactionItem'
        type: array
      mileage:
        description: Пробіг при транзакції (для авто)
        type: integer
      note:
        type: string
      photos:
        items:
          $ref: '#/definitions/models.TransactionPhoto'
        type: array
      receipt_img:
        description: Основне фото (legacy або прев'ю)
        type: string
      related_transaction:
        $ref: '#/definitions/models.Transaction'
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      transfer_related_id:
        description: Перекази (Transfer)
        type: string
      type:
        type: string
      updated_at:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.TransactionItem:
    properties:
      category_id:
        description: Категорія конкретного товару
        type: string
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      id:
        type: string
      is_synced:
        type: boolean
      name:
        type: string
      price_per_unit:
        type: integer
      quantity:
        type: integer
      total_amount:
        type: integer
      transaction_id:
        type: string
      updated_at:
        type: integer
    type: object
  models.TransactionPhoto:
    properties:
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      id:
        type: string
      is_synced:
        type: boolean
      path:
        type: string
      transaction_id:
        type: string
      updated_at:
        type: integer
    type: object
...
  models.User:
    properties:
      avatar_url:
        type: string
      base_currency:
        description: Налаштування користувача
        type: string
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      email:
        description: Додав unique
        type: string
      family:
        $ref: '#/definitions/models.Family'
      family_id:
        type: string
      id:
        type: string
      is_synced:
        type: boolean
      language:
        type: string
      name:
        type: string
      password_hash:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        description: Relations
      role_id:
        type: string
      theme:
        type: string
      updated_at:
        type: integer
    type: object
...
  repositories.TopStat:
    properties:
      currency:
        type: string
      metadata:
        type: string
      name:
        type: string
      total:
        type: integer
    type: object
  repositories.TrendStat:
    properties:
      currency:
        type: string
      date:
        type: string
      total:
        type: integer
    type: object
...
  services.DashboardData:
    properties:
      total_balance:
        type: integer
      total_expense:
        type: integer
      total_income:
        type: integer
    type: object
...
info:
  contact:
    name: Vlad Hrytsaiuk
  description: API for WeGaS Finance application.
  title: WeGaS Finance API
  version: "1.0"
paths:
...
  /dashboard/recent:
    get:
      consumes:
      - application/json
      description: Returns the most recent transactions for the dashboard.
      parameters:
      - description: Comma-separated account IDs
        in: query
        name: account_ids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Transaction'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get recent transactions
      tags:
      - Dashboard
  /dashboard/stats:
    get:
      consumes:
      - application/json
      description: Returns main dashboard statistics like total balance, total income,
        and total expenses.
      parameters:
      - description: Start date (timestamp)
        format: int64
        in: query
        name: from
        type: integer
      - description: End date (timestamp)
        format: int64
        in: query
        name: to
        type: integer
      - description: Comma-separated account IDs
        in: query
        name: account_ids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.DashboardData'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get dashboard statistics
      tags:
      - Dashboard
  /dashboard/top:
    get:
      consumes:
      - application/json
      description: Returns top entities (categories or counterparties) for income
        or expenses.
      parameters:
      - description: Flow type (income, expense)
        in: query
        name: type
        required: true
        type: string
      - description: Entity type (category, counterparty)
        in: query
        name: entity
        required: true
        type: string
      - description: Start date (timestamp)
        format: int64
        in: query
        name: from
        type: integer
      - description: End date (timestamp)
        format: int64
        in: query
        name: to
        type: integer
      - description: Comma-separated account IDs
        in: query
        name: account_ids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repositories.TopStat'
            type: array
        "400":
          description: Missing params
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get top statistics
      tags:
      - Dashboard
  /dashboard/trend:
    get:
      consumes:
      - application/json
      description: Returns daily aggregated income or expense trend over a period.
      parameters:
      - description: Flow type (income, expense)
        in: query
        name: type
        required: true
        type: string
      - description: Override target currency
        in: query
        name: currency
        type: string
      - description: Start date (timestamp)
        format: int64
        in: query
        name: from
        type: integer
      - description: End date (timestamp)
        format: int64
        in: query
        name: to
        type: integer
      - description: Comma-separated account IDs
        in: query
        name: account_ids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repositories.TrendStat'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get financial trend
      tags:
      - Dashboard
  /export/transactions:
    get:
      consumes:
      - application/json
      description: Returns a list of transactions based on filters, suitable for export
        (e.g., CSV/PDF generation on the frontend).
      parameters:
      - description: Start date (timestamp)
        format: int64
        in: query
        name: from
        type: integer
      - description: End date (timestamp)
        format: int64
        in: query
        name: to
        type: integer
      - description: Comma-separated account IDs
        in: query
        name: accountIds
        type: string
      - description: Comma-separated category IDs
        in: query
        name: categoryIds
        type: string
      - description: Comma-separated user IDs
        in: query
        name: userIds
        type: string
      - description: Comma-separated counterparty IDs
        in: query
        name: counterpartyIds
        type: string
      - collectionFormat: csv
        description: Transaction types
        in: query
        items:
          type: string
        name: type
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Transaction'
            type: array
        "400":
          description: Invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Export transactions
      tags:
      - Export
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"